server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9 


SQLAlchemy==2.0.30
alembic==1.13.1
PyMySQL==1.1.1 

firebase-admin==6.5.0
supabase==2.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0 


httpx==0.27.0
requests==2.31.0
transformers==4.40.0  

openai==1.30.1
google-generativeai==0.8.3  # Gemini API (Vision 대체용)
Pillow==10.4.0  # 이미지 처리 (Gemini Vision용)
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
langchain-chroma==0.1.3  # ChromaDB 벡터 스토어 (deprecated 경고 해결)
chromadb==0.5.0 # 벡터 데이터베이스


pydantic==2.7.1
beautifulsoup4==4.12.2supabase>=2.0.0
